+++ frontend/src/components/ContestHistoryTable.js
@@ -1,56 +1,30 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 import {
-  withStyles,
   makeStyles,
-  useTheme,
 } from '@material-ui/core/styles';
 import {
-  IconButton,
   Table,
   TableBody,
-  TableCell,
   TableContainer,
   TableFooter,
   TableRow,
   TablePagination,
   Paper,
 } from '@material-ui/core';
-import {
-  FirstPage,
-  KeyboardArrowLeft,
-  KeyboardArrowRight,
-  LastPage,
-} from '@material-ui/icons';
 
 import SortableTableHead, {
   getComparator,
   stableSort,
 } from './SortableTableHead';
-
-const StyledTableCell = withStyles((theme) => ({
-  head: {
-    backgroundColor: theme.palette.common.black,
-    color: theme.palette.common.white,
-  },
-  body: {
-    fontSize: 14,
-  },
-}))(TableCell);
-
-const StyledTableRow = withStyles((theme) => ({
-  root: {
-    '&:nth-of-type(odd)': {
-      backgroundColor: theme.palette.background.default,
-    },
-  },
-}))(TableRow);
+import {
+  StyledTableCell,
+  StyledTableRow,
+} from './StyledTable';
+import {
+  TablePaginationActions,
+} from './TablePagination';
 
 const useStyles = makeStyles((theme) => ({
-  root: {
-    flexShrink: 0,
-    marginLeft: theme.spacing(2.5),
-  },
   table: {
     minWidth: 650,
   },
@@ -69,81 +43,6 @@ const useStyles = makeStyles((theme) => ({
     width: 1,
   },
 }));
-
-// TODO: Refactoring as a component.
-function TablePaginationActions(props) {
-  const classes = useStyles();
-  const theme = useTheme();
-  const { count, page, rowsPerPage, onChangePage } = props;
-
-  const handleFirstPageButtonClick = (event) => {
-    onChangePage(event, 0);
-  };
-
-  const handleBackButtonClick = (event) => {
-    onChangePage(event, page - 1);
-  };
-
-  const handleNextButtonClick = (event) => {
-    onChangePage(event, page + 1);
-  };
-
-  const handleLastPageButtonClick = (event) => {
-    onChangePage(event, Math.max(0, Math.ceil(count / rowsPerPage) - 1));
-  };
-
-  return (
-    <div className={classes.root}>
-      <IconButton
-        onClick={handleFirstPageButtonClick}
-        disabled={page === 0}
-        aria-label="first page"
-      >
-        {theme.direction === 'rtl'
-          ? <LastPage />
-          : <FirstPage />
-        }
-      </IconButton>
-      <IconButton
-        onClick={handleBackButtonClick}
-        disabled={page === 0}
-        aria-label="previous page"
-      >
-        {theme.direction === 'rtl'
-          ? <KeyboardArrowRight />
-          : <KeyboardArrowLeft />
-        }
-      </IconButton>
-      <IconButton
-        onClick={handleNextButtonClick}
-        disabled={page >= Math.ceil(count / rowsPerPage) - 1}
-        aria-label="next page"
-      >
-        {theme.direction === 'rtl'
-          ? <KeyboardArrowLeft />
-          : <KeyboardArrowRight />
-        }
-      </IconButton>
-      <IconButton
-        onClick={handleLastPageButtonClick}
-        disabled={page >= Math.ceil(count / rowsPerPage) - 1}
-        aria-label="last page"
-      >
-        {theme.direction === 'rtl'
-          ? <FirstPage />
-          : <LastPage />
-        }
-      </IconButton>
-    </div>
-  );
-}
-
-TablePaginationActions.propTypes = {
-  count: PropTypes.number.isRequired,
-  onChangePage: PropTypes.func.isRequired,
-  page: PropTypes.number.isRequired,
-  rowsPerPage: PropTypes.number.isRequired,
-};
 
 export default function ContestHistoryTable({ data }) {
   const classes = useStyles();
