+++ frontend/src/components/StyledTable.js
@@ -1,0 +1,25 @@
+import {
+  withStyles,
+} from '@material-ui/core/styles';
+import {
+  TableCell,
+  TableRow,
+} from '@material-ui/core';
+
+export const StyledTableCell = withStyles((theme) => ({
+  head: {
+    backgroundColor: theme.palette.common.black,
+    color: theme.palette.common.white,
+  },
+  body: {
+    fontSize: 14,
+  },
+}))(TableCell);
+
+export const StyledTableRow = withStyles((theme) => ({
+  root: {
+    '&:nth-of-type(odd)': {
+      backgroundColor: theme.palette.background.default,
+    },
+  },
+}))(TableRow);
