+++ frontend/src/components/TablePagination.js
@@ -1,0 +1,113 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import {
+  makeStyles,
+  useTheme,
+} from '@material-ui/core/styles';
+import {
+  IconButton,
+} from '@material-ui/core';
+import {
+  FirstPage,
+  KeyboardArrowLeft,
+  KeyboardArrowRight,
+  LastPage,
+} from '@material-ui/icons';
+
+const useStyles = makeStyles((theme) => ({
+  root: {
+    flexShrink: 0,
+    marginLeft: theme.spacing(2.5),
+  },
+//   table: {
+//     minWidth: 650,
+//   },
+//   container: {
+//     maxHeight: 640,
+//   },
+//   visuallyHidden: {
+//     border: 0,
+//     clip: 'rect(0 0 0 0)',
+//     height: 1,
+//     margin: -1,
+//     overflow: 'hidden',
+//     padding: 0,
+//     position: 'absolute',
+//     top: 20,
+//     width: 1,
+//   },
+}));
+
+export function TablePaginationActions(props) {
+  const classes = useStyles();
+  const theme = useTheme();
+  const { count, page, rowsPerPage, onChangePage } = props;
+
+  const handleFirstPageButtonClick = (event) => {
+    onChangePage(event, 0);
+  };
+
+  const handleBackButtonClick = (event) => {
+    onChangePage(event, page - 1);
+  };
+
+  const handleNextButtonClick = (event) => {
+    onChangePage(event, page + 1);
+  };
+
+  const handleLastPageButtonClick = (event) => {
+    onChangePage(event, Math.max(0, Math.ceil(count / rowsPerPage) - 1));
+  };
+
+  return (
+    <div className={classes.root}>
+      <IconButton
+        onClick={handleFirstPageButtonClick}
+        disabled={page === 0}
+        aria-label="first page"
+      >
+        {theme.direction === 'rtl'
+          ? <LastPage />
+          : <FirstPage />
+        }
+      </IconButton>
+      <IconButton
+        onClick={handleBackButtonClick}
+        disabled={page === 0}
+        aria-label="previous page"
+      >
+        {theme.direction === 'rtl'
+          ? <KeyboardArrowRight />
+          : <KeyboardArrowLeft />
+        }
+      </IconButton>
+      <IconButton
+        onClick={handleNextButtonClick}
+        disabled={page >= Math.ceil(count / rowsPerPage) - 1}
+        aria-label="next page"
+      >
+        {theme.direction === 'rtl'
+          ? <KeyboardArrowLeft />
+          : <KeyboardArrowRight />
+        }
+      </IconButton>
+      <IconButton
+        onClick={handleLastPageButtonClick}
+        disabled={page >= Math.ceil(count / rowsPerPage) - 1}
+        aria-label="last page"
+      >
+        {theme.direction === 'rtl'
+          ? <FirstPage />
+          : <LastPage />
+        }
+      </IconButton>
+    </div>
+  );
+}
+
+TablePaginationActions.propTypes = {
+  count: PropTypes.number.isRequired,
+  onChangePage: PropTypes.func.isRequired,
+  page: PropTypes.number.isRequired,
+  rowsPerPage: PropTypes.number.isRequired,
+};
